# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.BrandCreateRequest:
    properties:
      description:
        maxLength: 255
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 100
        minLength: 2
        type: string
      slug:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - name
    - slug
    type: object
  models.BrandUpdateRequest:
    properties:
      description:
        maxLength: 255
        type: string
      is_active:
        type: boolean
      name:
        maxLength: 100
        minLength: 2
        type: string
      slug:
        maxLength: 100
        minLength: 2
        type: string
    type: object
...
  models.SwaggerBrand:
    description: Brand model for Swagger documentation
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      description:
        example: Apple Inc.
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      name:
        example: Apple
        type: string
      slug:
        example: apple
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
...
  response.Response:
    description: Standard API response structure
    properties:
      code:
        example: 200
        type: integer
      data: {}
      error: {}
      message:
        example: Operation successful
        type: string
      status:
        description: '"success" | "error"'
        example: success
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A TechStore e-commerce API built with Go
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TechStore API
  version: "1.0"
paths:
...
  /brands:
    get:
      consumes:
      - application/json
      description: Retrieve all brands
      produces:
      - application/json
      responses:
        "200":
          description: Brands retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.SwaggerBrand'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get all brands
      tags:
      - brands
    post:
      consumes:
      - application/json
      description: Create a new brand (Admin only)
      parameters:
      - description: Brand data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BrandCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Brand created successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SwaggerBrand'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create new brand
      tags:
      - brands
  /brands/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a brand (Admin only)
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Brand deleted successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Brand not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete brand
      tags:
      - brands
    get:
      consumes:
      - application/json
      description: Retrieve a specific brand by ID
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Brand retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SwaggerBrand'
              type: object
        "404":
          description: Brand not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get brand by ID
      tags:
      - brands
    put:
      consumes:
      - application/json
      description: Update brand information (Admin only)
      parameters:
      - description: Brand ID
        in: path
        name: id
        required: true
        type: string
      - description: Brand update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.BrandUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Brand updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.SwaggerBrand'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Brand not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update brand
      tags:
      - brands
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"